pyproject.toml
[project]
name = "question-compiler"
version = "0.1.0"
description = "This project compiles questions from images using OCR technology."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
]

[dependency-groups]
dev = [
    "marimo[mcp]>=0.17.0",
]
